UT(serde): separate deserializing using macros.
